+++ src/css/app.css
@@ -34,14 +34,17 @@ a-link {
   border: 1px solid rgba(0, 0, 0, 0.12);
   border-radius: 4px;
   background-color: #ffffff;
-  display: flex;
-  flex-direction: column;
   width: 518px;
   min-height: 593px;
   margin: 0 auto;
-  align-items: center;
   margin-top: 32px;
   padding: 40px;
+}
+
+.page {
+  align-items: center;
+  display: flex;
+  flex-direction: column;
 }
 
 .title {
@@ -89,11 +92,11 @@ a-link {
   color: white;
 }
 
-.page-container {
+.tab-container {
   margin-top: 52px;
 }
 
-.page {
+.tab {
   display: flex;
   flex-direction: column;
   gap: 48px;
